// tmp/src/app/model/import-devices-interface.ts
export interface ImportDevices {
  email: string;
  deviceType: string;
  deviceStatusLoad: number;
  moveDate: Date;
  invoiceProvider: string;
  invoiceVale: string;
  sapPlant: string;
  emailDeviceManager: string;
  calledCode: string;
  importFileBase64: string;
  fileName: string;
  managementId: string;
  groupName: string;
  groupId: string;
}