// tmp/src/app/shared/enums/statusLoad.enums.ts
export enum StatusLoadEnum {
  Aquisicao = 1,
  Devolucao = 2,
  Perda = 3,
  AtivacaoDeLicenca = 4,
  TransferenciaDeLocal = 5,
  TransferenciaDePropriedade = 6,
  EnvioAssistenciaTecnica = 7,
  RetornoAssistenciaTecnica = 8,
  RetornoAssistenciaTecnicaSubstituido = 9,
  RetornoAssistenciaTecnicaNovo = 10,
  Descarte = 11,
}